Adapted jnf test code to have 1 passing test. Yay.
